kg_construction/feature_extractor/requirements.txt
wheel
numpy
torch
torchvision
torchaudio
resampy
librosa
h5py
pillow
imageio